feat: put current workflow state in recoil
