+++ packages/twenty-front/src/modules/workflow/states/currentWorkflowState.ts
@@ -1,0 +1,15 @@
+import { FlowData } from '@/workflow/types/Workflow';
+import { createState } from 'twenty-ui';
+
+export const currentWorkflowState = createState<{
+  data: FlowData | undefined;
+  loading: boolean;
+  error: Error | undefined;
+}>({
+  key: 'currentWorkflowState',
+  defaultValue: {
+    data: undefined,
+    error: undefined,
+    loading: true,
+  },
+});
